requirements.txt
fastapi 
uvicorn 

pydantic 
pydantic-settings

python-dotenv 

openai
aiofiles 
python-Levenshtein